feat: rename package from orionmipt to ljv
