+++ src/main/java/ljv/Quote.java
@@ -1,4 +1,4 @@
-package orionmipt;
+package ljv;
 
 public class Quote {
