+++ src/test/java/ljv/Node.java
@@ -1,4 +1,4 @@
-package orionmipt;
+package ljv;
 
 class Node {
     String name;
